Conflicts:
	em/packages/eman2/protocol_boxing.py
	protocol/protocol.py
	web/pages/resources/css/project_content_style.css
	web/pages/templates/project_content.html

+++ utils/path.py
@@ -80,10 +80,11 @@ def joinExt(*extensions):
 def cleanPath(*paths):
     """Remove a list of paths, either folders or files"""
     for p in paths:
-        if isdir(p):
-            shutil.rmtree(p)
-        else:
-            os.remove(p)
+        if exists(p):
+            if isdir(p):
+                shutil.rmtree(p)
+            else:
+                os.remove(p)
             
 def makePath(*paths):
     """Create a list of paths if they don't exists"""
